remove unwrap for getting random formation, create objective and run module in interface for events
