+++ crates/thetawave_interface/src/audio.rs
@@ -3,6 +3,7 @@ use bevy_kira_audio::AudioTween;
 use serde::Deserialize;
 use strum_macros::Display;
 
+/// Background music types
 #[derive(Deserialize, Debug, Hash, PartialEq, Eq, Clone, Display)]
 pub enum BGMusicType {
     Game,
@@ -30,6 +31,7 @@ pub enum SoundEffectType {
     MegablastAbility,
 }
 
+/// Subtype of sound effect for collisions
 #[derive(Deserialize, Debug, Hash, PartialEq, Eq, Clone, Display, Default)]
 pub enum CollisionSoundType {
     Squishy,
@@ -39,9 +41,13 @@ pub enum CollisionSoundType {
 
 #[derive(Event, Default)]
 pub struct ChangeBackgroundMusicEvent {
+    /// Background music to change to, None will just stop the current music
     pub bg_music_type: Option<BGMusicType>,
+    /// Loop from a specific time in the track, None will not loop the track
     pub loop_from: Option<f64>,
+    /// Fade in the music cycling in (bg_music_type)
     pub fade_in_tween: Option<AudioTween>,
+    /// Fade out the music currently being played
     pub fade_out_tween: Option<AudioTween>,
 }
