+++ crates/thetawave_interface/src/lib.rs
@@ -2,5 +2,7 @@ pub mod audio;
 pub mod character_selection;
 pub mod game;
 pub mod health;
+pub mod objective;
+pub mod run;
 pub mod spawnable;
 pub mod states;
