+++ src/player/systems/mod.rs
@@ -3,15 +3,11 @@
 mod ability;
 mod attacks;
 mod movement;
-use crate::{
-    game::GameParametersResource,
-    misc::HealthComponent,
-    run::{RunDefeatType, RunEndEvent, RunOutcomeType},
-    spawnable::SpawnEffectEvent,
-};
+use crate::{game::GameParametersResource, misc::HealthComponent, spawnable::SpawnEffectEvent};
 
 use bevy::prelude::*;
 use thetawave_interface::audio::{PlaySoundEffectEvent, SoundEffectType};
+use thetawave_interface::run::{RunDefeatType, RunEndEvent, RunOutcomeType};
 use thetawave_interface::spawnable::EffectType;
 
 pub use self::ability::*;
